+++ workbench/x3/rule_defns/bench.tmp
@@ -3,19 +3,23 @@ method[1]=RULE2RHS_CTX_MI_ALL
 method[2]=RULE2RHS_CTX_MI_REC
 method[3]=RULE2RHS_GET_RHS_ALL
 method[4]=RULE2RHS_GET_RHS_REC
+method[5]=RULE2RHS_GET_RHS_CRTP
 rules	method	sample	user	system	elapsed	size_out	
-4	0	1	0.33	0.01	0.35	131696
-4	1	1	0.27	0.02	0.31	93136
-4	2	1	0.26	0.03	0.30	84416
-4	3	1	0.28	0.01	0.30	75176
-4	4	1	0.28	0.00	0.30	80768
-8	0	1	0.68	0.01	0.70	297696
-8	1	1	0.33	0.03	0.37	158936
-8	2	1	0.38	0.00	0.39	147536
-8	3	1	0.33	0.01	0.35	119928
-8	4	1	0.35	0.00	0.36	135128
-12	0	1	6.94	0.03	6.98	546176
-12	1	1	0.40	0.03	0.44	236800
-12	2	1	1.16	0.00	1.17	221288
-12	3	1	0.38	0.01	0.41	160672
-12	4	1	0.65	0.00	0.66	205904
+8	0	1	0.66	0.02	0.69	298408
+8	1	1	0.36	0.01	0.37	159440
+8	2	1	0.36	0.02	0.39	147904
+8	3	1	0.33	0.01	0.36	120384
+8	4	1	0.34	0.02	0.37	135488
+8	5	1	0.32	0.02	0.35	115064
+12	0	1	6.94	0.03	6.97	547336
+12	1	1	0.42	0.01	0.44	237560
+12	2	1	1.16	0.00	1.17	221824
+12	3	1	0.37	0.03	0.41	161352
+12	4	1	0.64	0.01	0.66	206424
+12	5	1	0.37	0.02	0.40	157232
+14	0	1	32.23	0.03	32.27	712528
+14	1	1	0.43	0.04	0.47	283296
+14	2	1	3.76	0.02	3.79	266328
+14	3	1	0.41	0.02	0.44	181864
+14	4	1	1.44	0.04	1.49	242344
+14	5	1	0.39	0.02	0.43	178344
